fix: address current issues preventing packaged app from successfully booting up
